added more functionlity to folders
